+++ libs/f-flow/src/f-canvas/models/f-canvas-change-event.ts
@@ -1,0 +1,8 @@
+import { IPoint } from '@foblex/2d';
+
+export class FCanvasChangeEvent {
+  constructor(
+    public readonly position: IPoint,
+    public readonly scale: number,
+  ) {}
+}