+++ libs/f-flow/src/f-canvas/utils/index.ts
@@ -1,0 +1,2 @@
+export * from './merge-f-canvas-config';
+export * from './with-f-canvas';