+++ libs/f-flow/src/provide-f-flow/f-flow-feature.ts
@@ -5,6 +5,7 @@ import { Provider } from '@angular/core';
  */
 export enum EFFlowFeatureKind {
   REFLOW_ON_RESIZE = 'reflow-on-resize',
+  CANVAS = 'canvas',
 }
 
 /**